Revert change to forked CGLIB code

+++ spring-core/src/main/java/org/springframework/cglib/proxy/Enhancer.java
@@ -1254,7 +1254,7 @@ public class Enhancer extends AbstractClassGenerator {
 			}
 
 			public int getOriginalModifiers(MethodInfo method) {
-				return (Integer) originalModifiers.get(method);
+				return ((Integer) originalModifiers.get(method)).intValue();
 			}
 
 			public int getIndex(MethodInfo method) {
